Match intent of Export#filename test to signature
The previous test's description did not match its use in the test;
it said the Method Under Test's implementation would use today's date
but instead used the value of the constructor's parameter.

Since the application never uses the second parameter in the constructor,
it looks like the parameter was only used for testability.

We can instead safely stub today's date using Timecop.

Also:

* Use shorter, Law of Demeter-friendly `Date.current`.
* Since the Method Under Test is `#filename`,
  invoke it during the Exercise phase,
  separated by blank lines.