Add FixFutureDatedEntriesTransition
Before we deployed #142, a bug existed where we could reply to an email
from last year and it would be dated in the future. This happened
because we always assumed the current year when dating entries. If an
email prompt from Dec 31, 2014 received a response on Jan 1, 2015, then
the entry would be dated Dec 31, 2015.

This transition fixes any entries that were created when the bug
existed. Any entry dated for the future will be moved one year into the
past (the correct date). There are 36 entries in production that will be
effected. The latest of these entries was created Jan 3 2015, just
before the fix was deployed.

```
> entries_in_the_future = Entry.where("date > ?", Date.today)

> entries_in_the_future.count
=> 36

> entries_in_the_future.map(&:created_at).last
=> Sat, 03 Jan 2015 00:00:59 UTC +00:00
```