Add patch for CVE-2020-5267
We are not actually vulnerable to this exploit because we don't inject
javascript, but including here just in case we decided to inject
javascript in the future. :)

This patch is included in later versions of Rails. We are adding the
patch here because we're not ready to update our Rails major version.